Move p2p_request file to P2P utils
