+++ elements/lisk-p2p/src/p2p.ts
@@ -39,7 +39,7 @@ import {
 	PeerLists,
 } from './utils/types';
 
-import { P2PRequest } from './p2p_request';
+import { P2PRequest } from './utils/p2p_request';
 export { P2PRequest };
 import {
 	selectPeersForConnection,
